requirements_test.txt
openapi-spec-validator
pycodestyle
flake8
mypy
coverage
pytest-aiohttp
